+++ demos/assets.c
@@ -393,7 +393,7 @@
 #include "assets/default/inc/images/visible.data"
 #include "assets/default/inc/images/warn.data"
 #include "assets/default/inc/images/zh.data"
-#endif/*WITH_STB_IMAGE*/
+#endif /*WITH_STB_IMAGE*/
 #ifdef WITH_VGCANVAS
 #include "assets/default/inc/images/ball.bsvg"
 #include "assets/default/inc/images/china.bsvg"
@@ -401,13 +401,13 @@
 #include "assets/default/inc/images/pointer.bsvg"
 #include "assets/default/inc/images/pointer_1.bsvg"
 #include "assets/default/inc/images/pointer_4.bsvg"
-#endif/*WITH_VGCANVAS*/
+#endif /*WITH_VGCANVAS*/
 #if defined(WITH_TRUETYPE_FONT)
 #include "assets/default/inc/fonts/default.res"
-#else/*WITH_TRUETYPE_FONT*/
+#else /*WITH_TRUETYPE_FONT*/
 #include "assets/default/inc/fonts/default.data"
-#endif/*WITH_TRUETYPE_FONT*/
-#endif/*WITH_FS_RES*/
+#endif /*WITH_TRUETYPE_FONT*/
+#endif /*WITH_FS_RES*/
 
 ret_t assets_init(void) {
   assets_manager_t* am = assets_manager();
@@ -693,7 +693,7 @@ ret_t assets_init(void) {
   assets_manager_add(am, image_pointer);
   assets_manager_add(am, image_pointer_1);
   assets_manager_add(am, image_pointer_4);
-#endif/*WITH_VGCANVAS*/
+#endif /*WITH_VGCANVAS*/
 #endif
 
   tk_init_assets();