+++ src/base/bitmap.h
@@ -24,116 +24,11 @@
 
 #include "tkc/color.h"
 #include "base/types_def.h"
+#include "base/graphic_buffer.h"
 
 BEGIN_C_DECLS
 
-struct _bitmap_t;
-typedef struct _bitmap_t bitmap_t;
-
 typedef ret_t (*bitmap_destroy_t)(bitmap_t* bitmap);
-
-/**
- * @enum bitmap_format_t
- * @prefix BITMAP_FMT_
- * @annotation ["scriptable"]
- * 位图格式常量定义。
- */
-typedef enum _bitmap_format_t {
-  /**
-   * @const BITMAP_FMT_NONE
-   * 无效格式。
-   */
-  BITMAP_FMT_NONE = 0,
-  /**
-   * @const BITMAP_FMT_RGBA8888
-   * 一个像素占用4个字节，RGBA占一个字节，按内存地址递增。
-   */
-  BITMAP_FMT_RGBA8888,
-  /**
-   * @const BITMAP_FMT_ABGR8888
-   * 一个像素占用4个字节，ABGR占一个字节，按内存地址递增。
-   */
-  BITMAP_FMT_ABGR8888,
-  /**
-   * @const BITMAP_FMT_BGRA8888
-   * 一个像素占用4个字节，BGRA占一个字节，按内存地址递增。
-   */
-  BITMAP_FMT_BGRA8888,
-  /**
-   * @const BITMAP_FMT_ARGB8888
-   * 一个像素占用4个字节，ARGB占一个字节，按内存地址递增。
-   */
-  BITMAP_FMT_ARGB8888,
-  /**
-   * @const BITMAP_FMT_RGB565
-   * 一个像素占用2个字节，RGB分别占用5,6,5位, 按内存地址递增。
-   */
-  BITMAP_FMT_RGB565,
-  /**
-   * @const BITMAP_FMT_BGR565
-   * 一个像素占用2个字节，BGR分别占用5,6,5位, 按内存地址递增。
-   */
-  BITMAP_FMT_BGR565,
-  /**
-   * @const BITMAP_FMT_RGB888
-   * 一个像素占用3个字节，RGB占一个字节，按内存地址递增。
-   */
-  BITMAP_FMT_RGB888,
-  /**
-   * @const BITMAP_FMT_BGR888
-   * 一个像素占用3个字节，RGB占一个字节，按内存地址递增。
-   */
-  BITMAP_FMT_BGR888,
-  /**
-   * @const BITMAP_FMT_GRAY
-   * 一个像素占用1个字节。
-   */
-  BITMAP_FMT_GRAY,
-  /**
-   * @const BITMAP_FMT_MONO
-   * 一个像素占用1比特。
-   */
-  BITMAP_FMT_MONO,
-} bitmap_format_t;
-
-/**
- * @enum bitmap_flag_t
- * @annotation ["scriptable"]
- * @prefix BITMAP_FLAG_
- * 位图标志常量定义。
- */
-typedef enum _bitmap_flag_t {
-  /**
-   * @const BITMAP_FLAG_NONE
-   * 无特殊标志。
-   */
-  BITMAP_FLAG_NONE = 0,
-  /**
-   * @const BITMAP_FLAG_OPAQUE
-   * 不透明图片。
-   */
-  BITMAP_FLAG_OPAQUE = 1,
-  /**
-   * @const BITMAP_FLAG_IMMUTABLE
-   * 图片内容不会变化。
-   */
-  BITMAP_FLAG_IMMUTABLE = 2,
-  /**
-   * @const BITMAP_FLAG_TEXTURE
-   * OpenGL Texture, bitmap的id是有效的texture id。
-   */
-  BITMAP_FLAG_TEXTURE = 4,
-  /**
-   * @const BITMAP_FLAG_CHANGED
-   * 如果是MUTABLE的图片，更新时需要设置此标志，底层可能会做特殊处理，比如更新图片到GPU。
-   */
-  BITMAP_FLAG_CHANGED = 8,
-  /**
-   * @const BITMAP_FLAG_PREMULTI_ALPHA
-   * 预乘alpha。
-   */
-  BITMAP_FLAG_PREMULTI_ALPHA = 16
-} bitmap_flag_t;
 
 /**
  * @class bitmap_t
@@ -178,11 +73,11 @@ struct _bitmap_t {
    */
   const char* name;
   /**
-   * @property {uint8_t*} data
+   * @property {graphic_buffer_t*} buffer
    * @annotation ["readable"]
    * 图片数据。
    */
-  const uint8_t* data;
+  graphic_buffer_t* buffer;
 
   bool_t is_gif;
   /*for gif begin*/
@@ -245,6 +140,36 @@ bitmap_t* bitmap_create_ex(uint32_t w, uint32_t h, uint32_t line_length, bitmap_
  * @return {uint32_t} 返回一个像素占用的字节数。
  */
 uint32_t bitmap_get_bpp(bitmap_t* bitmap);
+
+/**
+ * @method bitmap_lock_buffer_for_read
+ * 为读取数据而锁定bitmap的图片缓冲区。
+ *
+ * @param {bitmap_t*} bitmap bitmap对象。
+ *
+ * @return {uint8_t*} 返回缓存区的首地址。
+ */
+uint8_t* bitmap_lock_buffer_for_read(bitmap_t* bitmap);
+
+/**
+ * @method bitmap_lock_buffer_for_write
+ * 为修改数据而锁定bitmap的图片缓冲区。
+ *
+ * @param {bitmap_t*} bitmap bitmap对象。
+ *
+ * @return {uint8_t*} 返回缓存区的首地址。
+ */
+uint8_t* bitmap_lock_buffer_for_write(bitmap_t* bitmap);
+
+/**
+ * @method bitmap_unlock_buffer
+ * 解锁图像缓冲区。
+ *
+ * @param {bitmap_t*} bitmap bitmap对象。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t bitmap_unlock_buffer(bitmap_t* bitmap);
 
 /**
  * @method bitmap_get_pixel
@@ -427,10 +352,6 @@ typedef enum _image_draw_type_t {
    */
   IMAGE_DRAW_PATCH3_Y_SCALE_X
 } image_draw_type_t;
-
-#ifndef BITMAP_ALIGN_SIZE
-#define BITMAP_ALIGN_SIZE 32
-#endif /*BITMAP_ALIGN_SIZE*/
 
 /*private*/
 ret_t bitmap_alloc_data(bitmap_t* bitmap);