+++ src/base/dialog_highlighter.c
@@ -27,7 +27,7 @@ ret_t dialog_highlighter_clear_image(dialog_highlighter_t* h) {
     vg = canvas_get_vgcanvas(h->canvas);
     vgcanvas_destroy_fbo(vg, &(h->fbo));
     memset(&(h->fbo), 0x00, sizeof(h->fbo));
-  } else if (h->img.data != NULL) {
+  } else if (h->img.buffer != NULL) {
     bitmap_destroy(&(h->img));
     memset(&(h->img), 0x00, sizeof(h->img));
   }