+++ src/base/types_def.h
@@ -336,4 +336,114 @@ typedef struct _system_info_t system_info_t;
 #define WITHOUT_DIALOG_HIGHLIGHTER 1
 #endif /*AWTK_LITE*/
 
+/**
+ * @enum bitmap_format_t
+ * @prefix BITMAP_FMT_
+ * @annotation ["scriptable"]
+ * 位图格式常量定义。
+ */
+typedef enum _bitmap_format_t {
+  /**
+   * @const BITMAP_FMT_NONE
+   * 无效格式。
+   */
+  BITMAP_FMT_NONE = 0,
+  /**
+   * @const BITMAP_FMT_RGBA8888
+   * 一个像素占用4个字节，RGBA占一个字节，按内存地址递增。
+   */
+  BITMAP_FMT_RGBA8888,
+  /**
+   * @const BITMAP_FMT_ABGR8888
+   * 一个像素占用4个字节，ABGR占一个字节，按内存地址递增。
+   */
+  BITMAP_FMT_ABGR8888,
+  /**
+   * @const BITMAP_FMT_BGRA8888
+   * 一个像素占用4个字节，BGRA占一个字节，按内存地址递增。
+   */
+  BITMAP_FMT_BGRA8888,
+  /**
+   * @const BITMAP_FMT_ARGB8888
+   * 一个像素占用4个字节，ARGB占一个字节，按内存地址递增。
+   */
+  BITMAP_FMT_ARGB8888,
+  /**
+   * @const BITMAP_FMT_RGB565
+   * 一个像素占用2个字节，RGB分别占用5,6,5位, 按内存地址递增。
+   */
+  BITMAP_FMT_RGB565,
+  /**
+   * @const BITMAP_FMT_BGR565
+   * 一个像素占用2个字节，BGR分别占用5,6,5位, 按内存地址递增。
+   */
+  BITMAP_FMT_BGR565,
+  /**
+   * @const BITMAP_FMT_RGB888
+   * 一个像素占用3个字节，RGB占一个字节，按内存地址递增。
+   */
+  BITMAP_FMT_RGB888,
+  /**
+   * @const BITMAP_FMT_BGR888
+   * 一个像素占用3个字节，RGB占一个字节，按内存地址递增。
+   */
+  BITMAP_FMT_BGR888,
+  /**
+   * @const BITMAP_FMT_GRAY
+   * 一个像素占用1个字节。
+   */
+  BITMAP_FMT_GRAY,
+  /**
+   * @const BITMAP_FMT_MONO
+   * 一个像素占用1比特。
+   */
+  BITMAP_FMT_MONO,
+} bitmap_format_t;
+
+/**
+ * @enum bitmap_flag_t
+ * @annotation ["scriptable"]
+ * @prefix BITMAP_FLAG_
+ * 位图标志常量定义。
+ */
+typedef enum _bitmap_flag_t {
+  /**
+   * @const BITMAP_FLAG_NONE
+   * 无特殊标志。
+   */
+  BITMAP_FLAG_NONE = 0,
+  /**
+   * @const BITMAP_FLAG_OPAQUE
+   * 不透明图片。
+   */
+  BITMAP_FLAG_OPAQUE = 1,
+  /**
+   * @const BITMAP_FLAG_IMMUTABLE
+   * 图片内容不会变化。
+   */
+  BITMAP_FLAG_IMMUTABLE = 2,
+  /**
+   * @const BITMAP_FLAG_TEXTURE
+   * OpenGL Texture, bitmap的id是有效的texture id。
+   */
+  BITMAP_FLAG_TEXTURE = 4,
+  /**
+   * @const BITMAP_FLAG_CHANGED
+   * 如果是MUTABLE的图片，更新时需要设置此标志，底层可能会做特殊处理，比如更新图片到GPU。
+   */
+  BITMAP_FLAG_CHANGED = 8,
+  /**
+   * @const BITMAP_FLAG_PREMULTI_ALPHA
+   * 预乘alpha。
+   */
+  BITMAP_FLAG_PREMULTI_ALPHA = 16
+} bitmap_flag_t;
+
+#ifndef BITMAP_ALIGN_SIZE
+#define BITMAP_ALIGN_SIZE 32
+#endif /*BITMAP_ALIGN_SIZE*/
+
+struct _bitmap_t;
+typedef struct _bitmap_t bitmap_t;
+
 #endif /*TK_TYPES_DEF_H*/