+++ src/ext_widgets/mutable_image/mutable_image.c
@@ -75,7 +75,7 @@ ret_t mutable_image_on_paint_self(widget_t* widget, canvas_t* canvas) {
       }
     }
 
-    if (bitmap->data != NULL) {
+    if (bitmap->buffer != NULL) {
       rect_t src = rect_init(0, 0, bitmap->w, bitmap->h);
       rect_t dst = rect_init(0, 0, widget->w, widget->h);
       canvas_draw_image(canvas, bitmap, &src, &dst);